pyproject.toml
[project]
name = "recap"
version = "2.0.0"
description = "Auto-capture your work from Git, Claude Code, and more. Generate reports and sync to Tempo."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Wei-Fan Liao" }
]
keywords = ["jira", "tempo", "worklog", "recap", "automation", "git", "time-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Scheduling",
]

dependencies = [
    "requests>=2.31.0",
    "rich>=13.0.0",  # 美化 CLI 輸出
    "typer>=0.12.0",  # 現代 CLI 框架
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
openai = [
    "openai>=1.50.0",
]
gemini = [
    "google-generativeai>=0.8.0",
]
all-llm = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
]
outlook = [
    "msal>=1.24.0",
]
excel = [
    "openpyxl>=3.1.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "greenlet>=3.0.0",
]
gitlab = [
    "python-gitlab>=4.0.0",
]
pe = [
    "reflex>=0.8.24",
    "pandas>=2.3.0",
    "openpyxl>=3.1.0",
    "openai>=1.50.0",
    "xlrd>=2.0.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
recap = "recap.cli:app"

[project.urls]
Homepage = "https://github.com/anthropics/recap"
Repository = "https://github.com/anthropics/recap"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recap"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/recap"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
